implementasi ability token pada project

+++ app/Http/Controllers/ProductController.php
@@ -46,6 +46,7 @@ class ProductController extends Controller
      */
     public function store(ProductRequest $request)
     {
+        $this->authorize('if_moderator');
 
         try{
             if ($request->price < 10000) {
@@ -102,6 +103,8 @@ class ProductController extends Controller
      */
     public function update(ProductRequest $request, Product $product)
     {
+        $this->authorize('if_admin');
+
         $attributes = $request->toArray();
         $attributes['slug'] = Str::slug($request->name . '-' . time());
         $product->update($attributes);
@@ -121,6 +124,8 @@ class ProductController extends Controller
      */
     public function destroy(Product $product)
     {
+        $this->authorize('if_admin');
+
         $product->delete();
 
         return response()->json([
